+++ CHANGELOG.md
@@ -7,6 +7,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 * Make events work on some CSS transforms
 * Fix caching on float dimensions
+* Fix `mouseleave` event on stage.
 
 ## 4.0.18 - 2019-11-20